refactor configs into public and private modules
