+++ pyproject.toml
@@ -19,6 +19,7 @@ dependencies = [
     "click>=8.1.7",
     "requests>=2.32.3",
     "pandas>=2.2.3",
+    "ipython>=8.32.0",
 ]
 
 [project.optional-dependencies]
