+++ src/scripts/network_setup/test_ether_network_client.py
@@ -1,7 +1,8 @@
 import time
 
 from ether import ether, ether_start, ether_get
-from ether._internal._config import EtherConfig, EtherNetworkConfig, EtherInstanceConfig
+from ether._internal._config import _EtherInstanceConfig
+from ether.config import EtherConfig, EtherNetworkConfig
 from ether.utils import get_ether_logger
 
 class NetworkClient:
@@ -50,7 +51,7 @@ def run_client(server_host: str):
     
     config = EtherConfig(
         instances={
-            "network_test_client": EtherInstanceConfig(
+            "network_test_client": _EtherInstanceConfig(
                 class_path="tests.test_ether_network_client.NetworkClient",
                 kwargs={"ether_name": "network_test_client"}
             )
