+++ tests/test_ether_get.py
@@ -1,9 +1,10 @@
 import pytest
 from ether import ether
 from ether import ether_get
+from ether.config import EtherConfig
 from ether.utils import get_ether_logger
 from typing import Optional, List
-from ether._internal._config import EtherConfig, EtherInstanceConfig
+from ether._internal._config import _EtherInstanceConfig
 
 class DataService:
     def __init__(self):
@@ -51,7 +52,7 @@ def test_ether_get_with_params():
     # Create configuration
     config = EtherConfig(
         instances={
-            "data_service": EtherInstanceConfig(
+            "data_service": _EtherInstanceConfig(
                 class_path="test_ether_get.DataService",
                 kwargs={"ether_name": "data_service"}
             )
