+++ tests/test_ether_network.py
@@ -2,8 +2,9 @@ import time
 from typing import List
 import socket
 from ether import ether
-from ether._internal._config import EtherConfig, EtherInstanceConfig, EtherNetworkConfig
+from ether._internal._config import _EtherInstanceConfig
 from ether import ether_save, ether_get
+from ether.config import EtherConfig, EtherNetworkConfig
 from ether.utils import get_ether_logger
 
 def get_local_ip():
@@ -68,7 +69,7 @@ def test_ether_network_communication():
     server_config = EtherConfig(
         network=server_network,
         instances={
-            "network_test": EtherInstanceConfig(
+            "network_test": _EtherInstanceConfig(
                 class_path="test_ether_network.NetworkTestService",
                 kwargs={"ether_name": "network_test"}
             )
