+++ tests/test_ether_reqrep_advanced.py
@@ -2,8 +2,9 @@ import pytest
 import time
 from typing import List, Optional
 from ether import ether
-from ether._internal._config import EtherConfig, EtherInstanceConfig
+from ether._internal._config import _EtherInstanceConfig
 from ether import ether_get, ether_save  
+from ether.config import EtherConfig
 from ether.utils import get_ether_logger
 import zmq
 from ether.liaison import EtherInstanceLiaison
@@ -53,7 +54,7 @@ def test_heartbeat_and_reconnect():
     
     config = EtherConfig(
         instances={
-            "heartbeat_service": EtherInstanceConfig(
+            "heartbeat_service": _EtherInstanceConfig(
                 class_path="test_ether_reqrep_advanced.HeartbeatService",
                 kwargs={"name": "heartbeat_service"}
             )
